src/email_worker/requirements.txt
boto3
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
twilio